Closes #16 No error pupup on creating new branch (#198)
* Added check for to see if a repo is opened

When attempting to create a new branch a check is carried out to see if there is a repo open. If no repo is open a error message is shown.

* Added check for to see if a repo is opened

When attempting to create a new branch a check is carried out to see if there is a repo open. If no repo is open a error message is shown.

* Added check for to see if a repo is opened

When attempting to create a new branch a check is carried out to see if there is a repo open. If no repo is open a error message is shown.